Extract SpringWebXmlAnalyzer.
Avoids duplicate code for people that build analyzers for web.xml.
